Ex 4.5 Helper functions step03
