+++ tests/list_helper.test.js
@@ -1,4 +1,5 @@
 const listHelper = require('../utils/list_helper');
+const testHelper = require('../utils/test_helpers');
 
 test('dummy returns one', () => {
   const blogs = [];
@@ -7,77 +8,33 @@ test('dummy returns one', () => {
 });
 
 describe('total likes', () => {
-  const listWithOneBlog = [
-    {
-      _id: '5a422aa71b54a676234d17f8',
-      title: 'Go To Statement Considered Harmful',
-      author: 'Edsger W. Dijkstra',
-      url: 'http://www.u.arizona.edu/~rubinson/copyright_violations/Go_To_Considered_Harmful.html',
-      likes: 5,
-      __v: 0,
-    },
-  ];
-
-  const listWithSeveralBlogs = [
-    {
-      _id: '5a422a851b54a676234d17f7',
-      title: 'React patterns',
-      author: 'Michael Chan',
-      url: 'https://reactpatterns.com/',
-      likes: 7,
-      __v: 0,
-    },
-    {
-      _id: '5a422aa71b54a676234d17f8',
-      title: 'Go To Statement Considered Harmful',
-      author: 'Edsger W. Dijkstra',
-      url: 'http://www.u.arizona.edu/~rubinson/copyright_violations/Go_To_Considered_Harmful.html',
-      likes: 5,
-      __v: 0,
-    },
-    {
-      _id: '5a422b3a1b54a676234d17f9',
-      title: 'Canonical string reduction',
-      author: 'Edsger W. Dijkstra',
-      url: 'http://www.cs.utexas.edu/~EWD/transcriptions/EWD08xx/EWD808.html',
-      likes: 12,
-      __v: 0,
-    },
-    {
-      _id: '5a422b891b54a676234d17fa',
-      title: 'First class tests',
-      author: 'Robert C. Martin',
-      url: 'http://blog.cleancoder.com/uncle-bob/2017/05/05/TestDefinitions.htmll',
-      likes: 10,
-      __v: 0,
-    },
-    {
-      _id: '5a422ba71b54a676234d17fb',
-      title: 'TDD harms architecture',
-      author: 'Robert C. Martin',
-      url: 'http://blog.cleancoder.com/uncle-bob/2017/03/03/TDD-Harms-Architecture.html',
-      likes: 0,
-      __v: 0,
-    },
-    {
-      _id: '5a422bc61b54a676234d17fc',
-      title: 'Type wars',
-      author: 'Robert C. Martin',
-      url: 'http://blog.cleancoder.com/uncle-bob/2016/05/01/TypeWars.html',
-      likes: 2,
-      __v: 0,
-    },
-  ];
-
   test('Empty list, should return 0', () => {
     expect(listHelper.totalLikes([])).toBe(0);
   });
 
   test('Single blog, should return its likes', () => {
-    expect(listHelper.totalLikes(listWithOneBlog)).toBe(5);
+    expect(listHelper.totalLikes(testHelper.listWithOneBlog)).toBe(5);
   });
 
   test('List with several blogs, should return the sum of all blogs likes', () => {
-    expect(listHelper.totalLikes(listWithSeveralBlogs)).toBe(36);
+    expect(listHelper.totalLikes(testHelper.listWithSeveralBlogs)).toBe(36);
+  });
+});
+
+describe('favorite blog', () => {
+  test('Empty blog list should return null', () => {
+    expect(listHelper.favoriteBlog([])).toBeNull();
+  });
+
+  test('Single blog list, should return blog', () => {
+    expect(listHelper.favoriteBlog(testHelper.listWithOneBlog)).toEqual(
+      testHelper.listWithOneBlog[0]
+    );
+  });
+
+  test('List with several blogs, should return fav blog', () => {
+    expect(listHelper.favoriteBlog(testHelper.listWithSeveralBlogs)).toEqual(
+      testHelper.listWithSeveralBlogs[2]
+    );
   });
 });
